requirements.txt
PyGithub==1.29
PyYAML==3.12
